refactor fast platform a bit
